refactor: decompose large files and organize structure
