+++ app/src/main/java/com/example/weatherapp/data/model/City.kt
@@ -1,0 +1,8 @@
+package com.example.weatherapp.data.model
+
+data class City(
+    val name: String,
+    val latitude: Double,
+    val longitude: Double,
+    val country: String
+)
