+++ app/src/main/java/com/example/weatherapp/data/model/CityWeather.kt
@@ -1,0 +1,6 @@
+package com.example.weatherapp.data.model
+
+data class CityWeather(
+    val city: City,
+    val weather: WeatherData
+)
