+++ app/src/main/java/com/example/weatherapp/data/model/Constants.kt
@@ -1,0 +1,3 @@
+package com.example.weatherapp.data.model
+
+val cities = listOf("Moscow", "London", "Paris", "Izhevsk", "Los Angeles")
