+++ app/src/main/java/com/example/weatherapp/data/model/WeatherData.kt
@@ -1,0 +1,8 @@
+package com.example.weatherapp.data.model
+
+data class WeatherData(
+    val latitude: Double,
+    val longitude: Double,
+    val timezone: String,
+    val hourly: HourlyData
+)
