+++ app/src/main/java/com/example/weatherapp/ui/details/WeatherDetailsScreen.kt
@@ -1,0 +1,73 @@
+package com.example.weatherapp.ui.details
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.weatherapp.ui.viewmodel.WeatherViewModel
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun WeatherDetailsScreen(
+    viewModel: WeatherViewModel,
+    cityName: String,
+    onBackClick: () -> Unit
+) {
+    val cityWeather = viewModel.getCityWeather(cityName)
+
+    Scaffold(
+        topBar = {
+            TopAppBar(
+                title = {
+                    Text(
+                        text = cityName,
+                        fontSize = 20.sp,
+                        fontWeight = FontWeight.Bold
+                    )
+                },
+                navigationIcon = {
+                    IconButton(onClick = onBackClick) {
+                        Icon(
+                            imageVector = Icons.Default.ArrowBack,
+                            contentDescription = "Назад"
+                        )
+                    }
+                }
+            )
+        }
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(innerPadding)
+                .verticalScroll(rememberScrollState())
+                .padding(16.dp)
+        ) {
+            if (cityWeather != null) {
+                CurrentWeatherSection(cityWeather)
+                ThreeDayForecastSection(cityWeather)
+                HourlyForecastSection(cityWeather)
+            } else {
+                Text(
+                    text = "Данные о погоде для $cityName не найдены",
+                    color = MaterialTheme.colorScheme.error
+                )
+            }
+        }
+    }
+}
